text_lookup.py is a prototype of using sql in addition to an index to provide on-the-fly chunking and citation creation for text down to the paragraph and chunked section level. This allows for easy chunk expansion by the start and end index and for grabbing a range of arbitrary paragraphs.
Later to include a chunk_index position so that there is an option to go up or down a specified # of questions.

Added initial notebooks for tracking the ingestion process before converting to script.

There is about a 50% loss on these parsers and there are duplication issues in the data. <- Tomorrow me's problem. Cleaning these up will come post-troubleshooting.